requirements.txt
# Flask Portfolio Project Dependencies
# Core framework
Flask==3.1.1

# Production WSGI server
gunicorn==21.2.0

# Environment variables management
python-dotenv==1.1.0

# Flask dependencies (automatically installed with Flask)
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3

# Note: The following packages are part of Python standard library:
# - smtplib (email sending)
# - ssl (secure connections)
# - email.message (email formatting)
# - os (operating system interface)
